+++ demo/src/demo/components/layout.nim
@@ -1,0 +1,14 @@
+import karax/[karaxdsl, vdom]
+
+import navbar, footer
+
+
+proc render*(content: VNode): VNode =
+  buildHtml(tdiv):
+    navbar.render()
+
+    main:
+      content
+
+    footer.render()
+