+++ demo/src/demo/components/panel/spacer.nim
@@ -1,0 +1,16 @@
+import karax/[kbase, vstyles]
+
+import ../component
+
+
+type Spacer* = object of Component
+  flex*: Natural
+
+
+proc initSpacer*(flex: Natural, customStyle = style()): Spacer =
+  let defaultStyle = style {
+    StyleAttr.flex: kstring $flex
+  }
+
+  Spacer(flex: flex, style: defaultStyle.merge(customStyle))
+