+++ demo/src/demo/state.nim
@@ -1,0 +1,9 @@
+import pages
+
+
+var
+  currentPage*: Page
+
+const
+  topPanelHeight* = "40px"
+