+++ src/sauer.nim
@@ -5,7 +5,10 @@ import sauer/commands/[sauer, pages]
 
 const commands = {
   "init": sauer.init,
+  "make": sauer.make,
+  "serve": sauer.serve,
   "pages new": pages.new,
+  "pages list": pages.list
 }