+++ changelog.md
@@ -7,6 +7,12 @@
 -   [t]—test suite improvement
 -   [d]—docs improvement
 
+## 1.0.1 (May 12, 2023)
+
+-   [+] Add `sauer pages list` command.
+-   [+] Add `sauer make` command.
+-   [+] Add `sauer serve` command.
+
 
 ## 1.0.0 (March 7, 2023)